pyproject.toml
[project]
name = "pi-tk-flow-ui"
version = "0.1.0"
description = "TUI for pi-tk-flow ticket management"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "pi-tk-flow contributors" }
]
keywords = ["pi", "tk", "workflow", "tui", "textual"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Terminals",
    "Typing :: Typed",
]

dependencies = []

[project.optional-dependencies]
ui = [
    "textual>=0.47.0",
    "pyyaml>=6.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
tf-ui = "pi_tk_flow_ui.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["python/pi_tk_flow_ui"]

[tool.hatch.build]
include = [
    "python/pi_tk_flow_ui/*.py",
    "python/pi_tk_flow_ui/*.tcss",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["python"]
addopts = "-v --tb=short"
